requirements.txt
torch>=1.10
numpy>=1.20.1
mrcfile>=1.3.0
scipy>=1.6.2
matplotlib>=3.3.4
setuptools>=52.0.0
sklearn>=0.0
scikit-learn>=0.24.1
biopython>=1.79
tensorboard>=2.8.0
tqdm
umap-learn